ae-waiting-time/requirements.txt
astroid==2.0.4
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nose==1.3.7
protobuf==3.7.1
pylint==2.1.1
requests==2.20.1
six==1.11.0
urllib3==1.24.1
wrapt==1.10.11
tabula-py==2.0.4
tika==1.23
pandas==0.23.4
